+++ src/factory/BFastOptions.js
@@ -1,7 +1,0 @@
-class BFastOptions {
-    constructor() {
-
-    }
-}
-
-module.exports = BFastOptions;