+++ src/factory/index.js
@@ -1,5 +1,0 @@
-const ResourceFactory = require('./ResourceFactory');
-
-module.exports = {
-    resources: ResourceFactory,
-};